move arrow_conversions into client, fix error, don't use default::*
