+++ kernel/src/client/mod.rs
@@ -1,5 +1,8 @@
 //! module for clients that are optionally built into the kernel
 
+#[cfg(feature = "arrow-conversion")]
+pub mod arrow_conversion;
+
 #[cfg(any(feature = "default-client", feature = "sync-client"))]
 pub mod arrow_data;
