+++ kernel/src/lib.rs
@@ -57,14 +57,12 @@ pub use error::{DeltaResult, Error};
 pub use expressions::Expression;
 pub use table::Table;
 
-#[cfg(feature = "arrow-conversion")]
-pub mod arrow_conversion;
-
-#[cfg(any(feature = "default-client", feature = "sync-client"))]
+#[cfg(any(
+    feature = "default-client",
+    feature = "sync-client",
+    feature = "arrow-conversion"
+))]
 pub mod client;
-
-#[cfg(feature = "default-client")]
-pub use client::default::*;
 
 /// Delta table version is 8 byte unsigned int
 pub type Version = u64;
